Fix adapative pooling kernel size bug on Power
The functions that PyTorch provides for doing adaptive pooling
on the CPU compute the kernel size based off of the size of the
input and output matrices. Rounding differences on Power cause
the results to be slightly off for these computations. This
results in pooling kernels that go over the edge of the matrix
and pick up junk data. Use doubles rather than floats for more
consistent behavior.

This bug was discovered by seeing failing tests in
test/onnx/test_models.py. The result given by VGG models were
different when run on the CPU with PyTorch vs. ONNX.

+++ aten/src/ATen/native/AdaptiveAveragePooling.cpp
@@ -14,7 +14,7 @@ namespace {
   }
 
   inline int end_index(int a, int b, int c) {
-    return (int)std::ceil((float)((a + 1) * c) / b);
+    return (int)std::ceil((double)((a + 1) * c) / b);
   }
 
   template <typename scalar_t>
