+++ aten/src/ATen/native/AdaptiveMaxPooling2d.cpp
@@ -14,7 +14,7 @@ inline int start_index(int a, int b, int c) {
 }
 
 inline int end_index(int a, int b, int c) {
-  return (int)std::ceil((float)((a + 1) * c) / b);
+  return (int)std::ceil((double)((a + 1) * c) / b);
 }
 
 // #define START_IND(a,b,c) a * c / b
